# [AutoTVM] Avoid using RPC for LocalRunner