@comaniac 2020-03-06
> Did you test GPU tuning? There are runtime issues like we cannot do fork after initializing CUDA runtime

That's a good point. I'll find a GPU instance to test it.