@masahi 2020-03-06
AMDGPU also needs RPC runner for autotvm, if I remember correctly