@comaniac 2020-03-06
@merrymercy I tested the local runner on Nvidia T4 and observed the issue you mentioned, so I put the get device attributes logic to a separate process so that the master process will not access the GPU context at all.

On the other hand, there's still an issue. The tuning process may still encounter initialization errors after a number of trials. I need to further investigate the root cause.