@jmorrill 2020-03-06
@comaniac When I was working on the Windows RPC server, my first design was to run the trials all in the same process as Windows doesn't `fork(...)`.  It did indeed have a speed improvement.

After running CUDA autotune a while it looked like there was a GPU memory leak (some trial made it jump hundreds of megs).  I spent a long time looking for a leak in the C++ code, but couldn't find anything.  With my very limited understanding, it seemed as if the failed/slow cuda kernels it would test, were themselves leaking.  Only killing the process would free the GPU memory.

I abandoned the in-process running of trials and moved to a similar design as the Linux current implementation.