@tqchen 2020-03-06
One thing I noticed is that a new RPC server and tracker seems is created each time `set_task` is called https://github.com/apache/incubator-tvm/blob/master/python/tvm/autotvm/measure/measure_methods.py#L333

Depending on how frequent that is called, perhaps we should change the server and tracker creation to only  once at the Local runner starting time, which should reduce the overhead.

@comaniac @merrymercy