@comaniac 2020-03-07
> One thing I noticed about the old LocalRunner is that a new RPC server and tracker seems is created each time `set_task` is called https://github.com/apache/incubator-tvm/blob/master/python/tvm/autotvm/measure/measure_methods.py#L333
> 
> Depending on how frequent that is called, perhaps we should change the server and tracker creation to only once at the Local runner starting time, which should reduce the overhead.
> 
> @comaniac @merrymercy

`set_task` is called only once per task so the overhead should not be a problem, especially I usually only test one task. Instead, I worry the stability of launching RPC server/tracker from master process. @merrymercy previously also found that this may cause the hanging problem. Maybe @merrymercy could comment more details.